First version of the monitor
Basic functionality: takes a server and Azuracast station shortcode,
builds the SSE client, and runs it, calling the callback of your choice.
A sample "dump the formatted data" function is included, but doing
more interesting things like sending the info to Discord is possible.

Updates are sent every time there is an event; it is up to you to
verify that you're not repeating yourself.